+++ stac_fastapi/elasticsearch/stac_fastapi/elasticsearch/app.py
@@ -17,7 +17,6 @@ from stac_fastapi.core.extensions import QueryExtension
 from stac_fastapi.core.extensions.aggregation import (
     EsAggregationExtensionGetRequest,
     EsAggregationExtensionPostRequest,
-    EsAsyncAggregationClient,
 )
 from stac_fastapi.core.extensions.fields import FieldsExtension
 from stac_fastapi.core.rate_limit import setup_rate_limit
@@ -39,6 +38,7 @@ from stac_fastapi.extensions.core import (
     TransactionExtension,
 )
 from stac_fastapi.extensions.third_party import BulkTransactionExtension
+from stac_fastapi.sfeos_helpers.aggregation import EsAsyncBaseAggregationClient
 from stac_fastapi.sfeos_helpers.filter import EsAsyncBaseFiltersClient
 
 logging.basicConfig(level=logging.INFO)
@@ -60,7 +60,7 @@ filter_extension.conformance_classes.append(
 )
 
 aggregation_extension = AggregationExtension(
-    client=EsAsyncAggregationClient(
+    client=EsAsyncBaseAggregationClient(
         database=database_logic, session=session, settings=settings
     )
 )
